requirements.txt
appdirs==1.4.3
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.4.25
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==2.0.1
